6809 assembler

+++ asm6800.h
@@ -1,4 +1,4 @@
-unsigned assemble(unsigned char *mem, unsigned addr, char *buf);
+unsigned short assemble(unsigned char *mem, unsigned short addr, char *buf);
 
 void show_syms(FILE *f);
 void clr_syms(void);
